requirements.txt
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
pytest==7.4.0
pytest-cov==4.1.0
mlflow==2.5.0
google-cloud-storage==2.10.0
seaborn==0.12.2
matplotlib==3.7.2
joblib==1.3.1
pyyaml==6.0
